requirements.txt
# Core dependencies
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
ta>=0.11.0

# Telegram integration (Fase 3)
python-telegram-bot>=20.0

# Optional: Charts and visualization
matplotlib>=3.7.0
plotly>=5.17.0
Pillow>=10.0.0

# Utilities
colorama>=0.4.6
requests>=2.31.0

# ML Dependencies
lightgbm>=4.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
numba>=0.58.0
imbalanced-learn>=0.11.0
